Websocket: use `FT_BYTES` for payloads
`doc/README.dissector` says `FT_NONE` is for fields which don't have a
value. Replace our use of `FT_NONE` in the websocket dissector with
`FT_BYTES` (and `BASE_NO_DISPLAY_VALUE` so the GUI rendering is
unchanged).

Main use case for this change is for extracting the payload using
`tshark -T fields`, so the payload fields give their hex content instead
of just the number 1.