WSUG: tweak Preferences info in popup menus
In sections 6.2.2 and 6.2.3 (documenting the context menus for the
Packet List and Packet Details panes respectively), have the description
of _Protocol Preferences_ link to section 11.5 (Preferences), add a note
linking to section 11.4.1 (the Enabled Protocols dialog box) in case
they disable the preference in this way.

In section 11.5, document that the above-mentioned context menus allow
directly opening the protocol's prefs window, setting individual prefs,
or disabling the protocol -- again with a link to section 11.4.1.