GSMTAP: Add the header (up to type) when the dissector table is used
Add the items for the protocol, the version, the header length, and
the type to the tree in all cases, including when the dissector table
is used.

This prevents having a gap in the dissected bytes (and no gsmtap
field or protocol even though gsmtap appears in frame.protocols)
for cases like GSMTAP logging packets.