epan: Don't set a libgcrypt log handler >= 1.11.0
On Windows, we've been setting a log handler for libgcrypt in order
to quiet messages, in particular one in the slow_gatherer routine:

  NOTE: you should run 'diskperf -y' to enable the disk statistics

Setting a log handler was deprecated in libgcrypt 1.11.0 and issues
a deprecation warning when compiling. That release also stopped
issuing that log message in cases where running diskperf wouldn't
help because DeviceIoControl(..., IOCTL_DISK_PERFORMANCE, ...) failed:

https://github.com/gpg/libgcrypt/commit/35abf4d2eb582b78873aa324f6d02976788ffbbc

So we should be able to stop suppressing the log messages on that
version.


(cherry picked from commit 5b5e506b88576658d743c7afd18490b05672bd5f)

Co-authored-by: John Thacker <johnthacker@gmail.com>