Qt: Update protocol preferences menu title on language change
The context menus for the PacketList and ProtoTree are created
on demand in a context menu list. Create the protocol preferences
menus at the same time instead of having the menus be a class
member that is cleared each time. This ensures that the title is
the translated title at the time of creation.

We were never adding the protocol preferences list on Logray;
now we don't bother to create the menu and go through a loop
filling it if we're not going to show it.

Take the opportunity to update to new-style signals and slots
in one case. (The other case requires using Q_DECLARE_OPAQUE_POINTER
on some Qt versions.)

Move an include from the ProtoTree header to the source file,
since we don't need it in the header now.

Reduce the translator burden slightly by using the
ProtocolPreferencesMenu translation of "ProtocolPreferences" for
the top level menu in the PacketList.

Part of #19997. The conversation colorize menus still need doing.