pcap: use nanosecond resolution for hardware lcap files.
The resolution on Ixia hardware-capture lcap files appears to be higher
than 1 microsecond; assume it's 1 nanosecond.

Also:

Make a Boolean variable a bool rather than an int.  Rename it to
skip_ixia_extra, to clarify that it's for the Ixia formats.

Use a null buffer argument to wtap_read_bytes() to skip over the size
information.

Point to the issue for adding support for Ixia lcap files.

Ping #14073.