pcap: fix heuristics for various modified pcap formats.
Sadly, some people created modified versions of pcap format that use
*the standard pcap magic number* but that aren't readable by code that
reads the standard format, requiring heuristics to figure out which
format is being used.

The heuristics try to read the file in all formats that use the magic
number in question, and choose the format that has the fewest problems
when that's done.

The old heuristics were a bit too eager to treat damaged regular pcap
files as "Red Hat 6.1" pcap files, as they looked less damaged (or
perhaps not damaged at all) when read in the latter format.

Redo the heuristic tests by:

* reading the packets with special code that reads the packet record
headers one field at a time, applying "does this look wrong?" tests as
soon as we have read enough from the header to apply the test, and
treating a short read as a "does this look wrong?" answer of "yes, it
does";

* doing "does this look wrong?" tests of fields that are not part of the
standard pcap packet header but are part of modified pcap headers;

* attempting to "read" the packet data (skipping over it and discarding
the data) as long as the captured length isn't larger than the maximum
we support;

* not applying the heuristics to pcap file version numbers other than
2.4.

(And, yes, even though the patches to Linux libpcap that caused some of
these problematic file formats were done back in 1999, at least one
capture file in one of those formats was attached to a bug filed within
the past 2 years or so - see issue #19318 - so the ability to handle
those formats appears still to be useful to at least some users.)