cllog: fix handling of EOF when processing the header.
The loop that processes the header comments can either terminate on an
EOF or error on on seeing a non-comment line.  Set linep on each trip
through the loop, so that, if it's null, we know the loop was terminated
due to an error or EOF; otherwise, it was terminated by a non-comment
line, which is in the line buffer.  Only parse the line buffer if
there's a non-comment line in the buffer.

Include the test for a comment line in the loop test.


(cherry picked from commit df34522dbc1193cae23e8823a0e0bfd05188cab7)

Co-authored-by: Guy Harris <gharris@sonic.net>