file_wrappers: clarify what file_getsp() and file_gets() do.
Expand the comment before file_getsp(), and add one before file_gets(),
indicating what it does.

Change the variables in file_getsp() so that buf always points to the
beginning of the buffer and a curp variable points to the *current*
position in the buffer, rather than using buf as that pointer.