packaging: Use _docdir in RPM spec; fix __bindir on SUSE
SUSE for some reason defines their __cmake and __ninja macros to
depend on _bindir; _bindir depends on _prefix and is the location
where the binaries for the package being built will be installed.
This isn't the case with __make, which is just defined as /usr/bin/make

But just because we could be installing to a different prefix doesn't
mean we want to run CMake or Ninja out of the directories we're
installing to. We want to use the system package CMake and Ninja.

Also, the newest SUSE doesn't have _docdir set to _datadir/doc by default,
so use the _docdir macro (which exists on RHEL/Fedora, and presumably
all RPM distributions) for the document file locations.