CMake: Respect TEST_EXTRA_ARGS, run release suite when requested
With the transition to using pytest, the CMake variable TEST_EXTRA_ARGS
stopped being used; several of the CI jobs are passing it, to no effect.

Note that this only has an effect with the `ninja test` target.
Running pytest directly requires passing in parameters manually.

However, we also install pytest inside a venv on gitlab macOS runner
jobs. So use the test target on the macOS package jobs, and instead
pass the text arguments to pytest on the GitLab macOS runners.