Qt: Set the DisplayFilterExpression dialog parent in child windows
If the DisplayFilterExpresion window is being launched from a
DisplayFilterEdit in a window other than the main window (and thus
other than from the main DisplayFilterCombo), set the new window
to be WindowModal. This also (after 67daaa34213bb10d493fcf88dfc21)
properly sets the parent relationship.

At a minimum, properly setting the parent widget is necessary in order
for the DFE window to be usable if the spawning window is modal;
otherwise, the parentless DFE is unusuable while the modal window
(which has the line edit intended to hold the DFE result) is open.

Fix #20049