json+json_3gpp: Ignore insignificant whitespace in compact mode
We have already tokenized the JSON when json_key_lookup is called.

Rather than doing offset and length calculations using strlen of the
key in the use_compact case, look at the appropriate token to find
the offset and length to use. This has the advantage of correctly
ignoring insignificant whitespace.