HTTP/2: Show the request URI in other frames on the stream
Add the generated request full URI field when we have it to the
other frames in the same HTTP/2 stream, such as the response HEADERS
and DATA frames. For the (rare) case of PUSH PROMISE frames, show
requests, if available, for both the current stream and the promised
one. Similar to #15344, but for HTTP/2.