Qt: Let traffic tabs with no data be the last selected tab
Use the tabData from the tabBar for the current index to get the
proto ID, instead of getting it from the ATapDataModel. The latter
only works when there's an entry in the model.

This allows switching to an empty tab and then disabling the old
initial tab.

Improve the fixes for #19869, #18250