epan: Add some comments about Little Endian NTP time
NTP time is only defined as Big Endian, as its a Internet protocol.
What "NTP timestamp, but encoded Little Endian" means can be different
depending on whether it's interpreted as a 64 bit fixed-point integer,
or as a struct with two 32 bit values, one for the seconds and one
for the fraction. RFCs 1305 and 4330 describe the timestamp as the
former, whereas RFC 5905 (which obsoletes those two) describes it as
the latter.

Currently, we implement the little endian variant as a struct of
seconds and fraction, similar to other little endian time types - but
no dissector uses this. ERF timestamps are the other format, so perhaps
we should shift to this.

RTPS has a similar timestamp which is like the NTP timestamp, but
using the Unix Epoch. RTPS *does* support choice of endianness, and
explicitly defines the timestamp as a struct of two 32 bit values,
so we continue to use that.

[skip ci]