filesystem: Only remove namespace from progfile_dir for extcaps
Extcaps are in their own subdirectory, so trim_progfile_dir trims that
off in that case to return the common program file directory.
Extcaps are now in a namespaced subdirectory, so we trim that off first
before looking to see if the next directory name matches "extcap".

However, if it's not an extcap directory, then restore the supposed
namespaced subdirectory. It could be "wireshark" from some other path,
particularly from \Program Files\wireshark (which, since Win32
GetModuleFileName doesn't "correct" the case of paths passed in on
the command line, can also happen if Wireshark is installed to the
default location.)

Fix #20057