http: Support Microsoft Xpress Compression in WUSP responses
Support Microsoft Xpress Compression in the Windows Update Services
HTTP-based Client-Server Protocol according to [MS-WUSP]

https://learn.microsoft.com/en-us/openspecs/windows_protocols/ms-wusp/

The stream is segmented into blocks and encoded with the Plain LZ77
version of [MS-XCA] Xpress Compression Algorithm

We need to define the decompression pref in all cases, since LZ77
is always available and needs no external libraries. Note that there
was already an issue in the (probably very rare) case of having
Snappy support but not Zlib or Brotli.

Fix #15692