Use ISO 8601 in more places for absolute times
Add conversion to ISO 8601 time representation as an option
to abs_time_to_str_ex. This is already preferred for creating
display filters from a field. There are three other categories
of absolute time representation.

* absolute_val_to_repr in ftype-time.c determines what is used
for outputting to -T fields, JSON, and the PDML "show" field.

* proto_item_fill_display_label controls what is put in custom
columns (and thus CSV output) and the packet diagrams.

* proto_item_fill_label and proto_tree_add_item_ret_time_string
determine what is placed in the proto tree, and hence the packet
details and the PDML "showname" field.

Add a preference to control when to use the previous time format,
which is similar to that of asctime. The three cases above are
ordered in terms how likely a machine versus a human is to consume
the data. Thus, have the preference be an enum using levels (rather
than a more complicate set of full 2^N choices.) By default, use
ISO 8601 format for the first two cases mentioned above, and the
existing asctime like format for the last.

Fix #10220