prefs: Write enum names to the prefs file, list both
The main way for users to determine possible command line preferences
is through the tshark -G currentprefs or tshark -G defaultprefs
reports.

When enum preference values were initially split into having both short
names and long descriptions, with the latter shown in the GUI but the
former useful for entering at a command line, the descriptions were
continued to be written to the preferences file instead of the names,
for backwards compatibility. This was done "for now", twenty years ago.

Both the command line and the preferences file support entering
either the name or description for an enum value, but only the
descriptions have been written as choices in the preferences file,
making it difficult to find the name options without searching through
the source code.

Print both the list of value names and the list of value descriptions as
part of the preceding comments for each preference. That is, if they're
different; for a number of enums the two sets are the same, in which
case only print one list. The descriptions are useful in the comments
for a fuller description of each option; the names are useful so that
a user knows what to put on the command line (or shorter options when
manually editing the preference file.)

Switch to writing out the shorter name in the preference file itself.
That is, hopefully it's shorter; there's no actual validation of that;
probably we should do validation of the "command-line friendly" names
similar to that done for preference names, etc.

This somewhat reduces backwards compatibility for anyone who writes a
modern preference file and then tries to open it with Ethereal 0.10.4,
but I can live with it.

Ping #15441