rpm packaging: Move unversioned shared library files to -devel package
The unversioned shared library files (just symbolic links)
like libwireshark.so, libwiretap.so, and libwsutil.so
belong in the -devel package. They're only necessary for linking
when compiling, and not for running the program. This doesn't
make a tremendous amount of difference in our typical use case,
but makes it somewhat easier to have multiple versions installed.

See also:
https://docs.fedoraproject.org/en-US/packaging-guidelines/#_devel_packages