prefs: Validate enum names
enum_val_t preferences have a name field, which is "the name one would
use on the command line for the value." Validate that those are actually
command line friendly, and don't require quoting, etc. (They're all
treated case-insensitively.) Long descriptions are shown in the GUI
and can have any characters.

Change enum names that didn't conform to this. In all but one case
(WiMAX ASN Control Protocol) there are no changes to the description -
in that case, the name and description were apparently switched but
the short name used a space.

Partially revert a3ee94499aa1baf3183b0fad07fe56390a9cb7ab
so that we continue to write the descriptions to the preference file,
so that older versions of Wireshark will continue to be able to
read preference values that they supported. (Continue to list the
names as options, so that there's some reference for command line use.)
We can switch this to the name once 5.0 is the oldest supported version.

This means that new Wireshark will still be able to read any preference
files written by earlier Wireshark (with the exception of wimaxasncp
if that preference was set to a non default value).

In the vast majority of these cases, the name field was the same as
the description. In the few exceptions (e.g. ANSI MAP and ANSI TCAP),
scripts that used the (longer) name instead of the description for
manually setting the preference value could have problems. Otherwise,
since the command line accepts both names and descriptions, there is
no impact to existing scripts.

Ping #15441