bt-dht: decode BEP42 ip field with v6 address
This adds decoding of IPv6 addresses into the correct IPv6 field
instead of outputting it as a hex encoded string. The ip field appears
in the wild with v6 addresses, and BEP42 in combination with BEP32 makes
it clear that the encoding is the same IP + port encoding used for v4.