tshark: Do not require -G to be the first option
Do not require -G to be the first option.

Have the glossary reports (other then the "defaultprefs" report)
reflect the current configuration state. That means dumping the
reports after other relevant options are applied, for example the
-o, -d, and --[disable|enable]-[protocol|heuristic] options.

For the new modified version of the "fields" report that filters on
certain prefixes, change the syntax to "-G fields,<prefix>", similar
to the "-z" options. Add a similar option for the "elastic-mapping"
report as an alternative to the separate --elastic-mapping-filter
long option. E.g.,

tshark -G elastic-mapping,ip,udp,dns

tshark -G elastic-mapping --elastic-mapping-filter ip,udp,dns

are both acceptable. (Perhaps the latter should be deprecated?)

Fix #17924