capture: Allow aborting a read with pending packets
It's possible in certain cases for dumpcap to have written packets
that haven't been added to the packet list yet. Normally (especially
when saving) we wait for Wireshark to finish processing those packets.

Allow the Qt GUI to set the capture file state to FILE_READ_ABORTED
in cases where we want to close a file without reading the remaining
packets. Note we don't necessarily want to exit the application in
that case; we could be doing this to restart a capture.

This doesn't fully address the case where we're in the middle of
stopping the capture waiting for packets to be added, and the user
has changed their mind and wants to exit now.

Part of #19831