Art-Net: Don't optimize a single-user regex
We create the regex each time, so G_REGEX_OPTIMIZE doesn't help
much and might even hurt. There appears to be a bug in accessing
unitialized memory on partial matches in Glib or PCRE2 when
optimization is enabled that doesn't occur when it isn't.

Also add some quantifier limits to bound the length of the regex.

Fix #20041.