epan: Fake nodes with null field_info to ensure proper tree structure
Faking nodes by returning the parent node has the problem of
producing an inaccurate tree structure. This isn't a problem
when initially adding nodes, but makes it impossible to later
change the length of the parent node. proto_item_set_len cannot
determine if the passed in node parameter contains the actual
field whose length should be changed, or if it contains a
parent node because the field was faked (in which case the call
should be ignored.)

Similarly, other routines that examine the tree structure, like
proto_item_get_parent, don't work properly on non-visible trees.

Deal with the problem by creating tree nodes with null field_info
structures and returning those. That should continue to have the
optimization benefits but avoid producing an incorrect structure.

The API functions generally deal with null field_info in a node
correctly, but dissectors and taps that examine the protocol tree
directly may not.

Add a hfinfo member to the node with the const header_field_info,
which exists and is known even when the node is faked.

Remove the MATE workaround of making the tree always visible. Now
that #17877 is fixed, #19619 works without this.

Fix #17877. Fix #18503.