protocol hierarchy tap: Don't fake protocols
The current cli protocol hierarchy tap sometimes includes the "Frame"
protocol, depending on whether filters (including from outside the tap)
filter on something in the Frame protocol. That's because it doesn't
actually set the tree's "fake_protocols" to false, but instead relies
on immediate children of the proto tree root node not being faked.

Have the protocolhierarchy CLI tap set a flag so that fake protocols
is set to false when it is used. This makes the Frame protocol always
show up, and matches what the GUI does.