wsutil: Remove an unnecessary component from the Windows extcap path
!14774 / 8541041922 ensured that we have separate namespaces everywhere
for syscall+log and packet extcaps. It went a bit too far and added an
unnecessary subdirectory to the global extcap path on Windows.

Add missing "running in build directory" logic on Windows and use it
to build the proper extcap paths for both the build and installation
directories.

Fixes #20069