GTPv2: Don't loop through the IE Instance vals list with NULL tree
The IE Instance vals list has 250+ members. Since the keys are
four uint8_t, it could be made into a value_string_ext, or perhaps
into a epan scoped map lazily created the first time it's needed.

In any case, we don't need to search it when the tree is NULL, and
right now the proportional impact of performance is quite high when
the tree is NULL.