ENIP/CIP I/O: Register the ENIP encapsulation of CIP I/O to ENIP
Register the dissector for CIP I/O packets as carried by EtherNet/IP
(with its Common Packet Format) over UDP to the ENIP protocol.
That dissector always indicates EtherNet/IP, and users prefer a
proto_enip protocol item to be added to the tree in the initial
part of the dissector. (A proto_cipio item is added later, in the
actual CIP I/O data.)

Fix #19517