wslua: Use a 64 bit function for masks
Don't use a function that does a cast to a 32 bit function as
an intermediate step if we want a 64 bit output.

Note that, we need to go back and fix all these casting functions
so that they handle it better when lua_Integer is 64 bit (but never
unsigned, though at some point casting from unsigned to signed will
be officially ok now that C++20 and C23 have decreed that numbers
must be two's complement.) Though users who worry about that should
use our Uint64 type anyway.

Ping #20094