file_wrappers: Cast size_t to unsigned for MSVC
For other reasons, we limit our buffer size to unsigned. The
LZ4 functions need a pointer to a size_t. Cast the size_t
result back to a unsigned when adjusting the buffer size so
that MSVC is happy. (This is guaranteed not to overflow.)

Ping #20094


(cherry picked from commit 88b46e6ea8571695613bf49f6614f879bfd86e8e)

Co-authored-by: John Thacker <johnthacker@gmail.com>