IEEE802154: Fix Multipurpose frame parsing
The 802.15.4 dissector uses legacy frame security processing
for frames with the version 802.15.4-2003, but these version
checks should only be performed for the generic frame format
as Multipurpose frames using separate version numbers, and
they all include the security header.

Without this fix, Multipurpose frames with the security
header were reported as invalid. With the fix, the security
header is correctly displayed.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>