dfilter: Support directional double quotation marks
Support the left double quotation mark (U+201C) in UTF-8 encoding
as beginning a double quoted string literal, and the right double
quotation mark (U+201D) as ending one. This mostly occurs when
filter examples are cut and paste from web pages or word processing
documents that "helpfully" substitute curly quotes for straight ones.

String literals can be surrounded by either a pair of straight double
quotes or a matching left and right directional double quotes, but
not a mix of straight and curly double quotes.

While doing this, convert the patterns in the flex for dfilter from
octal to hex, which is likely a little easier for people to read
these days, and it's what people are more likely to see if they look
up UTF-8 characters.