CMake: Set Python3_FIND_REGISTRY NEVER on MSYS2
On Windows, by default, CMake FindPython3 will prefer Python
installed in locations mentioned in the registry in to locations
mentioned in PATH and other relevant environment variables.
On MSYS2, the PATH is set up to properly contain the correct
Python3 and we never want to use Python mentioned in the registry.

Do that and we can use FindPython3 instead of the deprecated
FindPythonInterp (which is removed by CMP0148, new in CMake 3.27).

This requires CMake 3.13, which is already our minimum.

See https://gitlab.kitware.com/cmake/cmake/-/issues/21658