androiddump: Don't pass a SOCKET as nfds to select on Win32
The Win32 winsock API indicates that it ignores the nfds parameter,
which "is only included for compatiblity with Berkeley sockets."
The parameter is an int, just like in UN*X.

However, the Win32 SOCKET type is 64 bits (even though in practice
only the 32 bits are used; this is just beause it's a pointer
apparently), causing MSVC to complain about possibly discarding data.
On Win32 can either cast the SOCKET to an int to be ignored,
or pass any integer literal (like 0) to be ignored.