tshark: Output the current list of columns with -G column-formats
Add a helper function to convert fmt_data to the string used in
preferences, and use it in several places.

Output the current column formats to the column-formats glossary
in addition to the example of how to output the default format.
Add somewhat more documentation about setting preferences at the
command line.

Ping #17332. Ping #15441.