MySQL: Handle embedded null bytes in queries
The MySQL/MariaDB protocol states that COM_QUERY and COM_STMT_PREPARE
use string<EOF> (to the end of the packet) to contain SQL queries,
because embedded NUL bytes are allowed:

https://dev.mysql.com/doc/dev/mysql-server/latest/page_protocol_com_query.html
https://dev.mysql.com/doc/dev/mysql-server/latest/page_protocol_com_stmt_prepare.html
https://mariadb.com/kb/en/com_query/
https://mariadb.com/kb/en/com_stmt_prepare/

Fix the MySQL-specific parts of #20101.

There's a separate issue affecting all of Wireshark involving proper
display and handling of string fields with embedded NULs.