Capture: don't retrieve both monitor mode and non monitor mode caps
Don't always retrieve the capabilities for non monitor mode and
monitor mode. It causes too many problems to put an interface into
monitor mode when not requested, at least on macOS. Some of the
related Linux issues (#20090 and #20100) were dealt with in another
manner, because they would cause problems even when actually
capturing in monitor mode.

This can perhaps be revisited if libpcap can find a way to retrieve
the link-layers supported when a device is in monitor-mode mode
without actually putting the device in monitor mode.

Fix #20051. Fix #20082.