capture: Disable interface monitor on Linux when getting caps
Changing the monitor mode settings of a device can create and/or
remove a temporary device under Linux. We don't want retrieving
the capabilities of an interface to trigger new interface list
refreshes, which lead to an infinite loop. That includes when
starting stats. For similar reasons, stop listening immediately
when beginning a capture.

Do this by telling the socket to stop listening for messages
temporarily while the retrieval is in progress. Actually removing
the socket would require more updates in the GUI; if we told the
GUI to stop processing we would have to wait to start listening
again when the event loop had processes all pending events on
the file descriptor.

Fix #20100. Fix #20090.