Qt: Don't refresh interfaces while capturing
If we get a request to refresh interfaces while capturing is
going, defer it until we're not longer capturing and process
the interface list then. Changing the interface list during a
capture isn't something we can really do much about; if we're
capturing on a single interface and it goes away, that will stop
capturing, at which point the interface list will be taken care
of.

Note that 89cc85b5efce579aeda3cfaeb0d74551b229acc4 in fixing #11176
for similar reasons disabled the menu item, so this is only from
interacing monitoring.

This prevents, among other things, the extcaps from being rediscovered
and their toolbars from being cleared. Fix #19854.

It also fixes some problems that occur when capturing on a device in
monitor mode.