bt-dht: flag duplicate and unordered keys
This adds an expert info flag for dictionary keys that don't appear in
sorted order as required by the bencode format defined in the BitTorrent
specification (BEP03). While the ordering is important for BitTorrent,
it isn't stictly required for the DHT protocol unless the receiver has a
strict parser. Because of this, and because it is common enough that
most implementations will handle it gracefully, the severity is set to
CHAT. ~0.2% of packets in the wild are expected to be flagged.

As a consequence of the implementation detecting duplicates in sorted
dicts becomes trivial so these cases are flagged as well. When testing
with a different implementation that could detect duplicates in out of
order dicts as well, no matches were found in a very large sample of
packets.