MMS: Add more null checks
Also don't register the protocol in tables a second time. We don't
need to call the handoff function when changing the preferences
because there's nothing it needs to do.

The files in #1342 crash when disabling the new preference without
checking if the private data is null.


(cherry picked from commit 906e55b8d588171d1995224a7cc012c2a13f3a21)

Co-authored-by: John Thacker <johnthacker@gmail.com>