dumpcap: Add a comment about capabilities and monitor mode
On Linux, if we capture on an mac80211 device and enable monitor mode
(which requires that libpcap is built with libnl support), that creates
a virtual device on the same wiphy that libpcap tries to delete when
capturing is done. Dropping privileges means that will fail with EPERM,
leaving the device hanging around to be deleted manually. (Or possibly
by something like NetworkManager.)

[skip ci]