dumpcap: Fix dumpcap capture child with LeakSanitizer standalone
The workaround used for ASan with included LSan needs also to be
used when compiling with LeakSanitizer standalone. For similar
reasons, the exit code when using LSan standalone (which is different
than the exit code that ASan with included LSan uses when it finds
a leak) also should be 0, at least when a capture child.

Fix up 416c99129f1c7af31aa24bfe0efe1dc63cf0791f