TLS, DTLS: Make dissector selection in the UAT more user-friendly
The user is expected to input a dissector name. Use the dissector
field UAT definition so that the GUI presents a list of valid
options to select from. Add a UAT definition for "dissector with
custom callback."

Make the ssl association info string more useful by including
dissector names (and descriptions) in the info message. Don't include
"TCP" or "UDP" in the message; if that's relevant, it will be included
in the dissector description.

Don't look up dissectors in the field/protocol filter name table;
dissectors are not protocols, and a dissector name may not be a
protocol name.

Fix #20106