Qt: 6.8 deprecations
QCheckBox::stateChanged was deprecated for QCheckBox::checkStateChanged,
which was introduced in Qt 6.7. The difference is that one has an
int as parameter, the other has the Qt::CheckState enum.

Also, Qt::TimeSpec was unified with QTimeZone in 6.5:
https://www.qt.io/blog/unifying-qttimespec-within-qtimezone
which overall makes things simpler, and the Qt::TimeSpec versions
are deprecated in 6.8

There are a small number of time related changes necessary for
qcustomplot as well, but that will be handled separately, as that
is upstream code.

The Qt documentation indicates that some of these are supposed to
be deprecated in Qt 6.9, but they are deprecated in the 6.8
release candidate. They might not be deprecated in the final Qt
6.8 packages, but it doesn't hurt to move to the new functions now.