Qt: Set HighDpiScaleFactorRoundingPolicy before creating the app
QGuiApplication::setHighDpiScaleFactorRoundingPolicy is a static
function that must be called before creating the application object,
according to the Qt documentation. Do so.

On recent Qt, there is a console warning if we attempt to set this
inside the main application after creating the application (and
possibly it has no effect.)

Note that we really only need to do this on Qt 5.14 and 5.15,
because the value we set it to is the default on Qt 6 or above.