QCustomPlot: Don't set -Werror on 3rd party code
QCustomPlot uses some aspects of Qt::TimeSpec which are deprecated
in Qt 6.8. Don't set -Werror on it as it's 3rd party code. We do
patch the code if it isn't fixed upstream in a timely manner, but
we don't need to break the build for using APIs deprecated by Qt.

Fixes building on Qt 6.8 for now.