PROXY: Use a CONVERSATION_PROXY type
We set the packet to use a different conversation when looking up
the conversation and attaching a dissector to the packet, so that
a conversation dissector set by, e.g., HTTP, doesn't interfere with
the conversation dissector set by PROXY. Use a conversation type for
proxy instead of the port type.

This avoids in a rare (illegal?) situation having an infinite loop
of the PROXY dissector calling itself (through the TCP dissector).

Fix #20111