CMake: Require CMake 3.16 or later
Qt says: "Qt projects are required to set version 3.16 or newer in the
cmake_minimum_required command to ensure a working set of policy behaviors."
https://doc.qt.io/qt-6/cmake-supported-cmake-versions.html

We haven't noticed any problems, but 3.16 is 5 years old, and all the
Linux distributions that currently build the latest sources come with
CMake 3.16 or later. Ubuntu 20.04 (focal), which goes end of standard
support in April 2025, has 3.16 (that focal is stuck on Qt 5.12 without
using a PPA is a bigger issue that may warrant dropping support before
the next major Wireshark release.)

Also update the example plugin minimum CMake version to match.