Clean up message boxes.
Use the text and informative text fields; don't use the detailed text
field, as not all platforms have it as a concept - on macOS, it shows up
as additional text after the main text, and on Windows and X11/Wayland,
it shows up as a concealed field that can be shown with a "show the
details" button.

Clean up some dialog contents.

Get rid of the old GTK+-derived hack with the message text being split
into primary and secondary text; we now set them as separate strings.