packaging: Add dftest and idl2wrs to -devel RPM
dftest is useful for development, add it to the "Development" component
(which is excluded from default targets).

idl2wrs is only used for development, so put it in the -devel package.
Move it to the "Development" component as well. (We still should rename
one of the two "idl2wrs" - there's the DCERPC one and the CORBA one.)

This doesn't affect packaging for other systems - it probably should
be shipped as part of a development package for Windows (#19976) or
macOS (#19010).

Ping #1199, #19976