WSDG: document how a protocol can "Follow Stream"
Increase the header level of the section _How to produce protocol
statistics (stats)_ to emphasize that it depends on information from the
previous section, _How to tap protocols_.

Add another sub-section under _How to tap protocols_ called _How to
follow protocol streams_ which gives a detailed overview of using
`register_follow_stream()` for a protocol. It is also a sub-section
since it also depends on having knowledge of taps.